布局资产适配
