+++ packages/gi-assets-basic/src/components/GroupBar/Component.less
@@ -1,19 +1,27 @@
-.gi-operator-bar{
+.gi-group-bar-wrapper{
+  position: relative;
+  width: 100%;
+  height:100%;
+  .gi-group-bar-sdk{
+    height:100%;
+  }
+}
+.gi-group-operator-bar{
   position:absolute;
   display: flex;
   align-items: center;
   justify-content: left;
-  border: 1px solid #eee;
   box-sizing: border-box;
   overflow-y: hidden;
   box-shadow: var(--box-shadow-base);
+  z-index:1;
   .gi-operator-bar-group{
     position: relative;
     display: flex;
     align-items: center;
     justify-content: left;
-    background-color: var(--background-color);
     height:100%;
+    background-color: var(--background-color);
     br{
       display: none;
       &+span{
