+++ packages/gi-assets-basic/src/components/GroupBar/OperatorItem/index.tsx
