+++ packages/gi-assets-basic/src/components/GroupBar/index.md
