+++ packages/gi-assets-basic/src/components/index.tsx
@@ -38,6 +38,7 @@ import ZoomOut from './ZoomOut';
 import ZoomStatus from './ZoomStatus';
 import RemoveNodeWithMenu from './RemoveNodeWithMenu';
 import OperatorBar from './OperatorBar';
+import GroupBar from './GroupBar';
 import SelectExchangeMenuItem from './SelectExchangeMenuItem';
 
 // 临时测试，后面要删掉，放到图谱业务资产库中
@@ -52,6 +53,7 @@ export {
   FitCenter,
   FitView,
   OperatorBar,
+  GroupBar,
   NodeLegend,
   MiniMap,
   CanvasSetting,
